requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
idna-ssl==1.1.0
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.6
pycparser==2.20
pyee==7.0.3
pylint==2.6.0
redis==3.5.3
requests==2.24.0
six==1.15.0
slackclient==2.8.2
slackeventsapi==2.2.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
